bps converted to tsx and adapted
